pyproject.toml
[tool.poetry]
name = "airflow-docker"
version = "0.1.0"
description = ""
authors = ["LAPTOP-VCGCOL4V <petercbull@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
apache-airflow = "^2.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
